*** empty log message ***

+++ src/sys/src/objects/tagm.c
@@ -1,4 +1,4 @@
-/*$Id: tagm.c,v 1.26 2000/05/05 22:14:00 balay Exp bsmith $*/
+/*$Id: tagm.c,v 1.27 2000/07/10 03:38:52 bsmith Exp balay $*/
 /*
       Some PETSc utilites
 */
@@ -41,7 +41,7 @@ int Petsc_DelTag(MPI_Comm comm,int keyval,void* attr_val,void* extra_state)
   int ierr;
 
   PetscFunctionBegin;
-  PLogInfo(0,"Petsc_DelTag:Deleting tag data in an MPI_Comm %d\n",(int)comm);
+  PLogInfo(0,"Petsc_DelTag:Deleting tag data in an MPI_Comm %d\n",(long)comm);
   ierr = PetscFree(attr_val);CHKERRQ(ierr);
   PetscFunctionReturn(MPI_SUCCESS);
 }
@@ -189,7 +189,7 @@ int PetscCommDuplicate_Private(MPI_Comm comm_in,MPI_Comm *comm_out,int* first_ta
     tagvalp[0] = *maxval;
     tagvalp[1] = 0;
     ierr       = MPI_Attr_put(*comm_out,Petsc_Tag_keyval,tagvalp);CHKERRQ(ierr);
-    PLogInfo(0,"PetscCommDuplicate_Private: Duplicating a communicator %d %d max tags = %d\n",(int)comm_in,(int)*comm_out,*maxval);
+    PLogInfo(0,"PetscCommDuplicate_Private: Duplicating a communicator %d %d max tags = %d\n",(long)comm_in,(long)*comm_out,*maxval);
   } else {
 #if defined(PETSC_USE_BOPT_g)
     int tag;
@@ -232,12 +232,9 @@ int PetscCommDestroy_Private(MPI_Comm *comm)
   }
   tagvalp[1]--;
   if (!tagvalp[1]) {
-    PLogInfo(0,"PetscCommDestroy_Private:Deleting MPI_Comm %d\n",(int)*comm);
+    PLogInfo(0,"PetscCommDestroy_Private:Deleting MPI_Comm %d\n",(long)*comm);
     ierr = MPI_Comm_free(comm);CHKERRQ(ierr);
   }
   PetscFunctionReturn(0);
 }
-
-
-
